requirements.txt
setuptools
pandas
numpy
cachetools
hubble_exchange==0.9.0rc10
python-binance 
python-dotenv
git+https://github.com/shubhamgoyal42/aio-binance-library
hyperliquid-python-sdk==0.1.19